feat: Provide separate starter
