--- dify/dify-spring-boot-starter/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,0 @@
-io.github.guoshiqiufeng.dify.config.DifyAutoConfiguration
