+++ dify/spring-boot-starter/build.gradle
@@ -1,0 +1,1 @@
+tasks.matching { it.group == 'publishing' }.each { it.enabled = false }
