+++ dify/spring-boot-starter/dify-spring-boot-starter-autoconfigure/build.gradle
@@ -1,0 +1,14 @@
+dependencies {
+    implementation project(":dify:dify-support:dify-support-chat")
+    implementation project(":dify:dify-support:dify-support-dataset")
+    implementation project(":dify:dify-support:dify-support-server")
+    implementation project(":dify:dify-support:dify-support-workflow")
+
+    implementation "org.springframework.boot:spring-boot-starter-data-redis"
+
+    implementation "org.springframework.boot:spring-boot-autoconfigure"
+    annotationProcessor "org.springframework.boot:spring-boot-starter-actuator"
+    annotationProcessor "org.springframework.boot:spring-boot-configuration-processor"
+
+    testImplementation "org.springframework.boot:spring-boot-starter-test"
+}
