+++ dify/spring-boot-starter/dify-spring-boot-starter-autoconfigure/src/main/java/io/github/guoshiqiufeng/dify/autoconfigure/DifyChatAutoConfiguration.java
@@ -1,0 +1,39 @@
+/*
+ * Copyright (c) 2025-2025, fubluesky (fubluesky@foxmail.com)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.github.guoshiqiufeng.dify.autoconfigure;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import io.github.guoshiqiufeng.dify.chat.impl.DifyChatDefaultImpl;
+import io.github.guoshiqiufeng.dify.core.config.DifyServerProperties;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+/**
+ * @author yanghq
+ * @version 1.0
+ * @since 2025/3/18 16:17
+ */
+@Configuration
+public class DifyChatAutoConfiguration {
+
+    @Bean
+    @ConditionalOnMissingBean({DifyChatDefaultImpl.class})
+    public DifyChatDefaultImpl difyChatHandler(DifyServerProperties difyServerProperties, ObjectMapper objectMapper) {
+        return new DifyChatDefaultImpl(difyServerProperties, objectMapper);
+    }
+
+}
