+++ dify/spring-boot-starter/dify-spring-boot-starter-autoconfigure/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,0 +1,5 @@
+io.github.guoshiqiufeng.dify.autoconfigure.DifyPropertiesAutoConfiguration
+io.github.guoshiqiufeng.dify.autoconfigure.DifyChatAutoConfiguration
+io.github.guoshiqiufeng.dify.chat.autoconfiguration.DifyDatasetAutoConfiguration
+io.github.guoshiqiufeng.dify.chat.autoconfiguration.DifyServerAutoConfiguration
+io.github.guoshiqiufeng.dify.chat.autoconfiguration.DifyWorkflowAutoConfiguration
