+++ dify/spring-boot-starter/dify-spring-boot-starter-server/build.gradle
@@ -1,0 +1,12 @@
+dependencies {
+    api project(":dify:spring-boot-starter:dify-spring-boot-starter-autoconfigure")
+
+    api project(":dify:dify-support:dify-support-server")
+
+    implementation "org.springframework.boot:spring-boot-starter-data-redis"
+
+    annotationProcessor "org.springframework.boot:spring-boot-starter-actuator"
+    annotationProcessor "org.springframework.boot:spring-boot-configuration-processor"
+
+    testImplementation "org.springframework.boot:spring-boot-starter-test"
+}
