+++ dify/spring-boot-starter/dify-spring-boot-starter-workflow/build.gradle
@@ -1,0 +1,10 @@
+dependencies {
+    api project(":dify:spring-boot-starter:dify-spring-boot-starter-autoconfigure")
+
+    api project(":dify:dify-support:dify-support-workflow")
+
+    annotationProcessor "org.springframework.boot:spring-boot-starter-actuator"
+    annotationProcessor "org.springframework.boot:spring-boot-configuration-processor"
+
+    testImplementation "org.springframework.boot:spring-boot-starter-test"
+}
